python-binance-wrapper/requirements.txt
flask
ccxt
pandas
requests
python-dotenv
